requirements.txt

tensorflow>=2.13.0
numpy
scikit-learn
matplotlib
pillow
streamlit
